### 2024-03-18 #240 @editorialbot
[REVIEW]: Self-Guided Decision Support Groundwater Modelling with Python
**Submitting author:** @rhugman (<a href="http://orcid.org/0000-0003-0891-3886">Rui Hugman</a>)
**Repository:** https://github.com/gmdsi/GMDSI_notebooks/
**Branch with paper.md** (empty if default branch): jose-review
**Version:** v1.0
