### 2024-05-06 #240 @incsanchezro
I am writing to provide feedback on your paper titled "Self-Guided Decision Support Groundwater Modelling with Python.". Following the JOSE guidelines for reviewers, I would like to confirm that I have no conflicts of interest in reviewing your work. Additionally, I affirm that I have adhered to the review code of conduct of JOSE as of April 2024.

Upon reviewing the paper and associated source code, I find that the overall quality of the learning framework meets essential criteria. I congratulate the authors for making the source code readily available on the specified GitHub repository. The inclusion of a generic plain-text LICENSE file reflects a commitment to open-source principles and provides clarity regarding usage rights. However, I recommend providing explicit identification of the software version as v1.0 to enhance clarity for users.

Furthermore, I appreciate the integrity with which authorship and contributions to the module have been managed. The primary author's significant and visible contributions demonstrate a profound commitment to the project's success and broader educational goals. While some authors may not have been actively engaged in the GitHub repository, their contributions are appropriately acknowledged in the written paper.

Overall, I recommend this paper for publication with attention to the comments listed below. I acknowledge the author's efforts in developing this valuable learning framework and look forward to seeing its continued evolution.

To the authors and JOSE editor, I extend my gratitude for the opportunity to review this work. Please do not hesitate to reach out if you require further clarification or feedback. While I will happily review the reviewed version of this publication.

**SPECIFIC COMMENTS AVAILABLE HERE**
1. >> https://github.com/openjournals/jose-reviews/issues/244

2.  >> https://github.com/openjournals/jose-reviews/issues/245

3. >> https://github.com/openjournals/jose-reviews/issues/246
