### 2024-05-06 #240 @incsanchezro
Dear @kls2177
Please find the reviewer comments for the paper "Self-Guided Decision Support Groundwater Modelling with Python" in the comment above ( >> https://github.com/openjournals/jose-reviews/issues/240#issuecomment-2096875833 ). Please accept my apologies for the delay.
