### 2024-05-06 b9ffe45
added pub plots to resp surf notebooks
