+++ paper/paper.bib
@@ -3,7 +3,8 @@
   title={Approaches to highly parameterized inversion: PEST++ version 5, a software suite for parameter estimation, uncertainty analysis, management optimization and sensitivity analysis},
   author={White, Jeremy T and Hunt, Randall J and Fienen, Michael N and Doherty, John E},
   year={2020},
-  institution={US Geological Survey Techniques and Methods 7-C26}
+  institution={US Geological Survey Techniques and Methods 7-C26},
+  doi = {10.3133/tm7C26}
 }
 
 @article{whitepyemu,
@@ -24,7 +25,8 @@ month = nov,
 pages = {217--228},
 title = {{A python framework for environmental model uncertainty analysis}},
 volume = {85},
-year = {2016}
+year = {2016},
+doi = {10.1016/j.envsoft.2016.08.017}
 }
 
 @techreport{White2020DR,
@@ -41,7 +43,8 @@ doi={10.5066/P9AUZMI7}
   volume={8},
   pages={50},
   year={2020},
-  publisher={Frontiers Media SA}
+  publisher={Frontiers Media SA},
+  doi = {10.3389/feart.2020.00050}
 }
 
 @article{white2021towards,
@@ -51,7 +54,8 @@ doi={10.5066/P9AUZMI7}
   volume={139},
   pages={105022},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/j.envsoft.2021.105022}
 }
 
 @article{white2022pestppmou,
@@ -61,7 +65,8 @@ doi={10.5066/P9AUZMI7}
   volume={149},
   pages={105316},
   year={2022},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/j.envsoft.2022.105316}
 }
 
 @software{Langevin_MODFLOW_6_Modular_2022,
@@ -90,7 +95,7 @@ journal = {Groundwater},
 volume = {54},
 number = {5},
 pages = {733-739},
-doi = {https://doi.org/10.1111/gwat.12413},
+doi = {10.1111/gwat.12413},
 url = {https://ngwa.onlinelibrary.wiley.com/doi/abs/10.1111/gwat.12413},
 eprint = {https://ngwa.onlinelibrary.wiley.com/doi/pdf/10.1111/gwat.12413},
 abstract = {Abstract Graphical user interfaces (GUIs) are commonly used to construct and postprocess numerical groundwater flow and transport models. Scripting model development with the programming language Python is presented here as an alternative approach. One advantage of Python is that there are many packages available to facilitate the model development process, including packages for plotting, array manipulation, optimization, and data analysis. For MODFLOW-based models, the FloPy package was developed by the authors to construct model input files, run the model, and read and plot simulation results. Use of Python with the available scientific packages and FloPy facilitates data exploration, alternative model evaluations, and model analyses that can be difficult to perform with GUIs. Furthermore, Python scripts are a complete, transparent, and repeatable record of the modeling process. The approach is introduced with a simple FloPy example to create and postprocess a MODFLOW model. A more complicated capture-fraction analysis with a real-world model is presented to demonstrate the types of analyses that can be performed using Python and FloPy.},
@@ -112,7 +117,8 @@ year = {2016}
   journal={Frontiers in Earth Science},
   volume={11},
   pages={1260581},
-  publisher={Frontiers}
+  publisher={Frontiers},
+  doi = {10.3389/978-2-8325-3581-3}
 }
 
 @article{fienen2022risk,
