+++ paper/paper.md
@@ -120,6 +120,12 @@ Each notebook in Part 0 is standalone and covers a unique topic. These include:
  - An introduction to the `pyemu` Python package that is used to complement and interface with PEST/PEST++.
  - Explanation of fundamental mathematical concepts that are relevant and will be encountered throughout the tutorial notebooks.
 
+Pre-requisites for Part 0 include a basic understanding of Python, Jupyter Notebooks, and MODFLOW 6. Familiarity with git is a bonus but not fundamental.
+
+After completing this module, learners should be able to:
+  - Describe the structure and components of the synthetic Freyberg model.
+  - Apply the `pyemu` Python package to interface with PEST/PEST++.
+  - Discuss the mathematical concepts that underpin the PEST/PEST++ software.
 
 ## Part 1: Introduction to PEST and the Gauss-Levenberg Marquardt Approach
 
@@ -133,6 +139,14 @@ Part 1 is designed to be accessible without strict sequential dependencies. Lear
 
  While Part 1 notebooks can be largely run in any order, the curriculum was initially designed to start with simple parameterization of a model and to build complexity intentionally throughout the progression of the sequence. The ramifications of simplification and the value of adding complexity are evaluated in the context of the performance of the model in forecasts made outside the parameter estimation conditions. This progression motivates the value of a highly-parameterized approach which is the starting point for many new projects, as explored in Part 2.
 
+Pre-requisites for Part 1 include a basic understanding of numerical groundwater modelling and familiarity with MODFLOW 6. Familiarity with Python and Jupyter Notebooks is assumed.
+
+After completing this module, learners should be able to:
+ - Describe the GLM approach to parameter estimation.
+ - Explain non-uniqueness, identifiability and equifinality in the context of decision-support modelling.
+ - Construct a PEST control file and interface files.
+ - Understand the effect of parameterization schemes on model predictive ability.
+ - Explain the difference between linear and non-linear uncertainy analysis approaches and the value of each.
 
 ## Part 2: Python-based Decision-Support Modelling Workflows
 Part 2 expands on the foundational knowledge gained in Part 1 and delves into advanced topics related to ensemble-based parameter estimation, uncertainty analysis and optimization methods. These advanced topics include management optimization and sequential data assimilation. This approach and these advanced topics assume a highly-parameterized approach, as motivated in Part 1. Topics are laid out in manner that reflects real-world workflows, with a focus on practical application of concepts and problem solving.
@@ -147,6 +161,14 @@ Part 2 is structured with a specific order for learners to follow to ensure a lo
 Each of these sequences comprises multiple notebooks to be executed in a specified order. They demonstrate how to execute the workflow, interpret results, and apply the concepts to real-world problems.
 
 In summary, the tutorial notebooks are organized to guide learners through a structured learning experience in the field of decision-support groundwater modelling. Part 0 provides foundational knowledge, while Parts 1 and 2 offer progressively advanced content. The authors attest that it is ideal to work through Parts 1 and 2 in their entirety, referring back to Part 0 for additional background. However, this amount of content requires a significant time commitment so, practically, many users will start with Part 2 and, hopefully, be able to apply the concepts to a problem of their own as they progress. Over time, referring back through Part 1 will provide a deeper understanding of some concepts and techniques taken for granted in the highly-parameterized, largely ensemble-based approaches of Part 2.
+
+Pre-requisites for Part 2 include a basic understanding of PEST/PEST++ and the PEST interface, as well as familiarity with the Freyberg model. Familiarity with Python and Jupyter Notebooks is assumed.
+
+After completing this module, learners should be able to:
+ - Construct a high-dimensional PEST control file and interface files.
[…]
