### 2024-05-08 da2f049
add learning objectives to README
