### 2024-05-08 7105b44
add some acronyms to improve clarity
