+++ paper/paper.bib
@@ -110,6 +110,18 @@ year = {2016}
   year={1994},
   publisher={Institution of Engineers, Australia Barton, ACT}
 }
+@book{Anderson2015,
+author = "Anderson, Mary P. and Woessner, William W.  and Hunt, Randall J. ",
+title = "Applied Groundwater Modeling",
+publisher = "Academic Press",
+edition = "Second",
+address = "San Diego",
+year = "2015",
+pages = "443 - 491",
+isbn = "978-0-12-058103-0",
+doi = "http://dx.doi.org/10.1016/B978-0-08-091638-5.00010-9"
+}
+
 
 @article{white11rapid,
   title={Rapid, Reproducible, and Robust Environmental Modeling for Decision Support: Worked Examples and Open-Source Software Tools},
@@ -153,3 +165,14 @@ year = {2016}
   year={2019},
   url={https://hbr.org/2019/10/where-companies-go-wrong-with-learning-and-development}
 }
+
+@ARTICLE{jupyter,
+  author={Granger, Brian E. and Pérez, Fernando},
+  journal={Computing in Science & Engineering}, 
+  title={Jupyter: Thinking and Storytelling With Code and Data}, 
+  year={2021},
+  volume={23},
+  number={2},
+  pages={7-14},
+  keywords={Open source software;Scientific computing;Machine learning;Data science;Open source software;Meteorology},
+  doi={10.1109/MCSE.2021.3059263}}
