+++ paper/paper.md
@@ -8,28 +8,28 @@ tags:
   - uncertainty analysis
 authors:
   - name: Rui T. Hugman
-    orcid: 0000-0003-0891-3886
-    affiliation: 1 
+.  orcid: 0000-0003-0891-3886
+.  affiliation: 1 
   - name: Jeremy T. White
-    orcid: 0000-0002-4950-1469
-    affiliation: 1 
+.  orcid: 0000-0002-4950-1469
+.  affiliation: 1 
   - name: Mike Fienen
-    orcid: 0000-0002-7756-4651
-    affiliation: "2" # (Multiple affiliations must be quoted)
+.  orcid: 0000-0002-7756-4651
+.  affiliation: "2" # (Multiple affiliations must be quoted)
   - name: Brioch Hemmings
-    orcid: 0000-0001-6311-8450
-    affiliation: "3" # (Multiple affiliations must be quoted)
+.  orcid: 0000-0001-6311-8450
+.  affiliation: "3" # (Multiple affiliations must be quoted)
   - name: Katie Markovitch
-    orcid: 0000-0002-4455-8255
-    affiliation: 1 
+.  orcid: 0000-0002-4455-8255
+.  affiliation: 1 
 
 affiliations:
  - name: INTERA Geosciences, Perth, Western Australia, Australia
-   index: 1
+. index: 1
  - name: U.S. Geological Survey, Upper Midwest Water Science Center, Madison, WI USA
-   index: 2
+. index: 2
  - name: Wairakei Research Centre, GNS Science, Taupō, New Zealand
-   index: 3
+. index: 3
 date: 20 September 2023
 bibliography: paper.bib
 
@@ -50,7 +50,7 @@ bibliography: paper.bib
 # format.
 # 
 # For a quick reference, the following citation commands can be used:
-# - `@author:2001`  ->  "Author et al. (2001)"
+# - `@author:2001. -. "Author et al. (2001)"
 # - `[@author:2001]` -> "(Author et al., 2001)"
 # - `[@author1:2001; @author2:2001]` -> "(Author1 et al., 2001; Author2 et al., 2002)"
 # 
@@ -62,18 +62,18 @@ bibliography: paper.bib
 
 # Summary
 
-The [GMDSI tutorial notebooks repository](https://github.com/gmdsi/GMDSI_notebooks/tree/main) provides learners with a comprehensive set of tutorials for self-guided training on decision-support groundwater modelling using Python-based tools. Although targeted at groundwater modelling, they are based around model-agnostic tools and readily transferable to other environmental modelling workflows. The tutorials are divided into three parts. The first covers fundamental theoretical concepts. These are intended as background reading for reference on an as-needed basis. Tutorials in the second part introduce learners to some of the core concepts parameter estimation in a groundwater modelling context, as well as providing a gentle introduction to the ``PEST``, ``PEST++`` and ``pyEMU`` software. Lastly, the third part demonstrates how to implement highly-parameterized applied decision-support modelling workflows. Their aim is to provide examples of both “how to use” the software as well as “how to think” about using the software. 
[…]
