+++ paper/peer_review/klj_review.md
@@ -1,0 +1,31 @@
+## General comments
+- Overall, this is a great presentation/summary of the notebooks and their utility, and I think the evaluation of them as a teaching tool is well done. Would be great to gather more data on that in the future!
+- Although the first paragraph of intro needs a little work, the rest of the intro does a great job of highlighting the issue that these notebooks help address.
+- There are a few locations where there are double spaces between sentences. Those should probably be changed to single?
+- It's good that you provide the context of how the notebooks might likely be used, like lines 191-200.
+
+## Line-specific comments/suggestions
+
+- 12: double space after "...parts."
+- 17: Unclear what "Their" refers to here. Switch to "The tutorials aim to provide..."?
+- 24-26: This first sentence is a bit dense, feels jargony. Because this is the first sentence of the article, I recommend simplifying language a little bit.
+- 28: "widely acknowledged." I agree, but citation needed?
+- 29: "assimilating data" might need to be explained a bit more. Reader might not know what "assimilation" means in this context.
+- 43: "use" instead of "uptake"?
+- 51: replace "can be manipulated using text files" with "text files as inputs"
+- 52: replace "that can be automatically extracted" with "are machine-readable"
+- 76: Could also mention that many GUIs are not open-source, reducing accessiblity to those who can afford.
+- 88: Is this true? I would tend to believe. Are anecdotal thoughts okay for this article? I think removing this sentence could be done, starting instead with the sentence on line 89 and replacing "this approach" with "Python packages like FloPy and pyEMU for model and PEST++ setup" 
+- 94: first use of USGS, write out.
+- 95: "power of jupyter notebooks" I think Jupyter is capitalized, and you maybe should explain what notebooks facilitate (code and explainations together). And cite Jupyter?
+- 98: first time mentioning MATLAB and R, are they discussed in the GMDSI notebooks? If not, I think you could remove.
+- 112: This connects to my comment for line 95, I think you need to be more clear right away on what makes the notebooks so useful.
+- 134: no hyphen in GitHub
+- 144: "visualizations" maybe?
+- 146: the use of dashes for the list folowing the colon is confusing, I thought something was missing. Were these supposed to be bulleted? If not, then I would recommend something other than dashes if it's going to be a long list as shown. 
+- 160: are these also supposed to be bullets?
+- 181: same comment as above
+- 209: Love the Discussions! maybe highlight that this can continue to be used well after the "experience"?
+- 212: figure link broken?
+- 230: This could be a good place to mention that the GitHub Discussions can serve as a virtual version of the in-person discussions/questions?
+- 238: missing "." after "U.S"
