+++ paper/paper.md
@@ -116,16 +116,16 @@ The tutorial notebooks are structured into three main parts:
 Part 0 serves as the foundation, providing essential background material. Learners are encouraged to reference notebooks in Part 0 to polish their understanding of concepts they encounter in Parts 1 and 2. Part 0 is not intended to be a comprehensive resource for all background material, but rather to establish a solid understanding of the basics. The explanations of mathematical concepts are intended to be accessible through visualization and descriptions related to everyday concepts and modelling concepts. 
 
 Each notebook in Part 0 is standalone and covers a unique topic. These include:
- -Introduction to a synthetic model known as the "Freyberg" model. This model is used as a consistent example throughout the tutorial exercises, allowing learners to apply concepts in a practical context.
- -An introduction to the `pyemu` Python package that is used to complement and interface with PEST/PEST++.
- -Explanation of fundamental mathematical concepts that are relevant and will be encountered throughout the tutorial notebooks.
+ * Introduction to a synthetic model known as the "Freyberg" model. This model is used as a consistent example throughout the tutorial exercises, allowing learners to apply concepts in a practical context.
+ * An introduction to the `pyemu` Python package that is used to complement and interface with PEST/PEST++.
+ * Explanation of fundamental mathematical concepts that are relevant and will be encountered throughout the tutorial notebooks.
 
 Pre-requisites for Part 0 include a basic understanding of Python, Jupyter Notebooks, and MODFLOW 6. Familiarity with git is a bonus but not fundamental.
 
 After completing this module, learners should be able to:
-  - Describe the structure and components of the synthetic Freyberg model.
-  - Apply the `pyemu` Python package to interface with PEST/PEST++.
-  - Discuss the mathematical concepts that underpin the PEST/PEST++ software.
+  * Describe the structure and components of the synthetic Freyberg model.
+  * Apply the `pyemu` Python package to interface with PEST/PEST++.
+  * Discuss the mathematical concepts that underpin the PEST/PEST++ software.
 
 ## Part 1: Introduction to PEST and the Gauss-Levenberg Marquardt Approach
 
@@ -142,21 +142,21 @@ Part 1 is designed to be accessible without strict sequential dependencies. Lear
 Pre-requisites for Part 1 include a basic understanding of numerical groundwater modelling and familiarity with MODFLOW 6. Familiarity with Python and Jupyter Notebooks is assumed.
 
 After completing this module, learners should be able to:
- - Describe the GLM approach to parameter estimation.
- - Explain non-uniqueness, identifiability and equifinality in the context of decision-support modelling.
- - Construct a PEST control file and interface files.
- - Understand the effect of parameterization schemes on model predictive ability.
- - Explain the difference between linear and non-linear uncertainy analysis approaches and the value of each.
+ * Describe the GLM approach to parameter estimation.
+ * Explain non-uniqueness, identifiability and equifinality in the context of decision-support modelling.
+ * Construct a PEST control file and interface files.
+ * Understand the effect of parameterization schemes on model predictive ability.
+ * Explain the difference between linear and non-linear uncertainy analysis approaches and the value of each.
 
 ## Part 2: Python-based Decision-Support Modelling Workflows
 Part 2 expands on the foundational knowledge gained in Part 1 and delves into advanced topics related to ensemble-based parameter estimation, uncertainty analysis and optimization methods. These advanced topics include management optimization and sequential data assimilation. This approach and these advanced topics assume a highly-parameterized approach, as motivated in Part 1. Topics are laid out in manner that reflects real-world workflows, with a focus on practical application of concepts and problem solving.
 
[…]
