+++ paper/paper.md
@@ -184,4 +184,7 @@ Open-ended feedback from the participants was generally positive and also includ
 
 The tutorials were originally developed with support from the US Geological Survey (USGS) and support from USGS continues through the HyTest training project. Continued development and support is funded by the Groundwater Modelling Decision Support Initiative (GMDSI). GMDSI is jointly funded by BHP and Rio Tinto. We thank Dr. John Doherty for his tireless and pioneering efforts starting `PEST` and continuing to innovate and Dr. Randall Hunt for his leadership in `PEST` and `PEST++` applications and development and contributions to the initial curriculum for this material and the early version of the notebooks. Kalle Jahn (USGS) and REVIEWER FROM JOURNAL? made reviews that improved this manuscript. We finally thank users and stress-testers for their valuable feedback and continued community contributions to the repository.
 
+# Disclaimer 
+Any use of trade, firm, or product names is for descriptive purposes only and does not imply endorsement by the US Government.
+
 # References
