+++ paper/paper.md
@@ -182,7 +182,7 @@ Open-ended feedback from the participants was generally positive and also includ
 
 # Acknowledgements
 
-The tutorials were originally developed with support from the US Geological Survey (USGS) and support from USGS continues through the HyTest training project. Continued development and support is funded by the Groundwater Modelling Decision Support Initiative (GMDSI). GMDSI is jointly funded by BHP and Rio Tinto. We thank Dr. John Doherty for his tireless and pioneering efforts starting `PEST` and continuing to innovate and Dr. Randall Hunt for his leadership in `PEST` and `PEST++` applications and development and contributions to the initial curriculum for this material and the early version of the notebooks. Kalle Jahn (USGS) and REVIEWER FROM JOURNAL? made reviews that improved this manuscript. We finally thank users and stress-testers for their valuable feedback and continued community contributions to the repository.
+The tutorials were originally developed with support from the US Geological Survey (USGS) and support from USGS continues through the HyTest training project. Continued development and support is funded by the Groundwater Modelling Decision Support Initiative (GMDSI). GMDSI is jointly funded by BHP and Rio Tinto. We thank Dr. John Doherty for his tireless and pioneering efforts starting `PEST` and continuing to innovate and Dr. Randall Hunt for his leadership in `PEST` and `PEST++` applications and development and contributions to the initial curriculum for this material and the early version of the notebooks. Kalle Jahn (USGS), [Ines Rodriguez](https://github.com/incsanchezro) and [REVIEWER2](https://github.com/codyalbertross) who made reviews that improved this manuscript. We finally thank users and stress-testers for their valuable feedback and continued community contributions to the repository.
 
 # Disclaimer 
 Any use of trade, firm, or product names is for descriptive purposes only and does not imply endorsement by the US Government.
