### 2024-05-09 96f76b8
fix paper render
