+++ paper/paper.md
@@ -116,47 +116,47 @@ The tutorial notebooks are structured into three main parts:
 Part 0 serves as the foundation, providing essential background material. Learners are encouraged to reference notebooks in Part 0 to polish their understanding of concepts they encounter in Parts 1 and 2. Part 0 is not intended to be a comprehensive resource for all background material, but rather to establish a solid understanding of the basics. The explanations of mathematical concepts are intended to be accessible through visualization and descriptions related to everyday concepts and modelling concepts. 
 
 Each notebook in Part 0 is standalone and covers a unique topic. These include:
- * Introduction to a synthetic model known as the "Freyberg" model. This model is used as a consistent example throughout the tutorial exercises, allowing learners to apply concepts in a practical context.
- * An introduction to the `pyemu` Python package that is used to complement and interface with PEST/PEST++.
- * Explanation of fundamental mathematical concepts that are relevant and will be encountered throughout the tutorial notebooks.
+ - Introduction to a synthetic model known as the "Freyberg" model. This model is used as a consistent example throughout the tutorial exercises, allowing learners to apply concepts in a practical context.
+ - An introduction to the `pyemu` Python package that is used to complement and interface with PEST/PEST++.
+ - Explanation of fundamental mathematical concepts that are relevant and will be encountered throughout the tutorial notebooks.
 
 Pre-requisites for Part 0 include a basic understanding of Python, Jupyter Notebooks, and MODFLOW 6. Familiarity with git is a bonus but not fundamental.
 
 After completing this module, learners should be able to:
-  * Describe the structure and components of the synthetic Freyberg model.
-  * Apply the `pyemu` Python package to interface with PEST/PEST++.
-  * Discuss the mathematical concepts that underpin the PEST/PEST++ software.
+  - Describe the structure and components of the synthetic Freyberg model.
+  - Apply the `pyemu` Python package to interface with PEST/PEST++.
+  - Discuss the mathematical concepts that underpin the PEST/PEST++ software.
 
 ## Part 1: Introduction to PEST and the Gauss-Levenberg Marquardt Approach
 
 Part 1 focuses on the Gauss-Levenberg Marquardt (GLM) approach to parameter estimation and associated uncertainty analysis in a groundwater modelling context. This was the foundation of the ``PEST`` software for multiple decades and the theory continues to resonate through newer techniques.
 
 Part 1 is designed to be accessible without strict sequential dependencies. Learners have the flexibility to explore its contents in any order that suits their preferences or needs. These include:
- -Introduction to concepts such as non-uniqueness, identifiability, and equifinality.
- -Introduction to the PEST control file and the PEST/PEST++ interface.
- -Exploring the challenges of parameterization schemes on predictive ability, as well as how to mitigate them.
- -Introducing first-order second-moment (FOSM) and prior Monte Carlo uncertainty analysis approaches.
+ - Introduction to concepts such as non-uniqueness, identifiability, and equifinality.
+ - Introduction to the PEST control file and the PEST/PEST++ interface.
+ - Exploring the challenges of parameterization schemes on predictive ability, as well as how to mitigate them.
+ - Introducing first-order second-moment (FOSM) and prior Monte Carlo uncertainty analysis approaches.
 
  While Part 1 notebooks can be largely run in any order, the curriculum was initially designed to start with simple parameterization of a model and to build complexity intentionally throughout the progression of the sequence. The ramifications of simplification and the value of adding complexity are evaluated in the context of the performance of the model in forecasts made outside the parameter estimation conditions. This progression motivates the value of a highly-parameterized approach which is the starting point for many new projects, as explored in Part 2.
 
[…]
