+++ paper/paper.md
@@ -67,11 +67,13 @@ A key advantage to using notebooks in this context is that the workflows describ
 
 # Statement of Need
 
+Managing the environment and natural resources is often faced with taking decisions on actions that trade off between economic, social, and environmental outcomes. In a groundwater management context, they may include decisions on such things as how much to pump from a well without impacting a nearby stream? or how to design a well-field to maximize yield while minimizing the risk of saltwater intrusion? These decisions are often complex and rely on the use of numerical models to simulate the natural system and evaluate the outcomes of different management actions [@kelly2013]. 
+
 Effective environmental management is often aided by using models. Such models face the challenges of limited data and the inherent simplifications of the complex natural world necessary to make them. This translates to uncertainties in critical decision-making predictions. Transparency is critical in reporting and mitigating these uncertainties, especially when significant risks accompany management outcomes. The significance of uncertainty quantification (UQ) and parameter estimation (PE) in environmental modeling for decision support is widely acknowledged [@Anderson2015]. UQ provides estimates of outcome uncertainty, while PE reduces this uncertainty through assimilating data, using a formal comparison of model outputs with observation data to inform model parameter values. 
 
-Implementing highly-parameterized UQ and PE in real-world modeling can be challenging due to both theoretical complexity and practical logistics. Limited project time and funding also often hinder their application. Open-source software such as ``PEST`` [@pest] and ``PEST++`` [@whitepestpp] provide tools for undertaking UQ and PE analyses. However, the steep learning curve associated with their use and the lack of user-friendly training materials have been a barrier to uptake.
+Implementing highly-parameterized UQ and PE in real-world modeling can be challenging due to both theoretical complexity and practical logistics. Limited project time and funding also often hinder their application. Open-source software such as ``PEST`` [@pest] and ``PEST++`` [@whitepestpp] provide tools for undertaking UQ and PE analyses. They have been employed to tackle a diverse range of real-world decision-support problems, ranging from risk-based well-head protection [@fienen2022risk], groundwater contaminant system remediation design [@FIENEN2024], evaluating the impact of groundwater use on river baseflow [@Foster2021],  managed aquiifer recharge schem design to protect against seawater intrusion [@Standen2022], optimizing vinyard irrigation [@KNOWLING2023108225], and many others. 
 
-There is a growing demand within the environmental modelling community for transparent, reproducible, and accountable modeling processes, driven by the need for increased credibility and rigor in computational science and environmental simulation [@white11rapid; @fienen2016hess]. While some script-based tools enhance the reproducibility of forward model construction [@flopy], they often overlook UQ and PE analyses. In decision-support scenarios, these analyses are equally vital for robust model deployment as the forward model itself. 
+However, the steep learning curve associated with their use and the lack of user-friendly training materials have been a barrier to uptake. There is a growing demand within the environmental modelling community for transparent, reproducible, and accountable modeling processes, driven by the need for increased credibility and rigor in computational science and environmental simulation [@white11rapid; @fienen2016hess]. While some script-based tools enhance the reproducibility of forward model construction [@flopy], they often overlook UQ and PE analyses. In decision-support scenarios, these analyses are equally vital for robust model deployment as the forward model itself. 
 
[…]
