+++ paper/paper.bib
@@ -176,3 +176,80 @@ doi = "http://dx.doi.org/10.1016/B978-0-08-091638-5.00010-9"
   pages={7-14},
   keywords={Open source software;Scientific computing;Machine learning;Data science;Open source software;Meteorology},
   doi={10.1109/MCSE.2021.3059263}}
+  
+ @article{kelly2013,
+  title = {Selecting among five common modelling approaches for integrated environmental assessment and management},
+  journal = {Environmental Modelling & Software},
+  volume = {47},
+  pages = {159-181},
+  year = {2013},
+  issn = {1364-8152},
+  doi = {10.1016/j.envsoft.2013.05.005},
+  url = {https://www.sciencedirect.com/science/article/pii/S1364815213001151},
+  author = {Rebecca A. {Kelly (Letcher)} and Anthony J. Jakeman and Olivier Barreteau and Mark E. Borsuk and Sondoss ElSawah and Serena H. Hamilton and Hans Jørgen Henriksen and Sakari Kuikka and Holger R. Maier and Andrea Emilio Rizzoli and Hedwig {van Delden} and Alexey A. Voinov},
+  keywords = {Integrated assessment, System dynamics, Bayesian network, Coupled component model, Agent-based model, Knowledge-based model},
+  abstract = {The design and implementation of effective environmental policies need to be informed by a holistic understanding of the system processes (biophysical, social and economic), their complex interactions, and how they respond to various changes. Models, integrating different system processes into a unified framework, are seen as useful tools to help analyse alternatives with stakeholders, assess their outcomes, and communicate results in a transparent way. This paper reviews five common approaches or model types that have the capacity to integrate knowledge by developing models that can accommodate multiple issues, values, scales and uncertainty considerations, as well as facilitate stakeholder engagement. The approaches considered are: systems dynamics, Bayesian networks, coupled component models, agent-based models and knowledge-based models (also referred to as expert systems). We start by discussing several considerations in model development, such as the purpose of model building, the availability of qualitative versus quantitative data for model specification, the level of spatio-temporal detail required, and treatment of uncertainty. These considerations and a review of applications are then used to develop a framework that aims to assist modellers and model users in the choice of an appropriate modelling approach for their integrated assessment applications and that enables more effective learning in interdisciplinary settings.}
+  }
+  
+@article{FIENEN2024,
+  title = {Comparing single and multiple objective constrained optimization algorithms for tuning a groundwater remediation system},
+  journal = {Environmental Modelling & Software},
+  volume = {173},
+  pages = {105952},
+  year = {2024},
+  issn = {1364-8152},
+  doi = {https://doi.org/10.1016/j.envsoft.2024.105952},
+  url = {https://www.sciencedirect.com/science/article/pii/S1364815224000136},
+  author = {Michael N. Fienen and Nicholas T. Corson-Dosch and Kalle L. Jahn and Jeremy T. White},
+  keywords = {Constrained optimization, Groundwater remediation, Numerical modeling, Environmental modeling},
[…]
