+++ paper/paper.md
@@ -40,7 +40,7 @@ bibliography: paper.bib
 # Include a “Statement of Need” section, explaining how the submitted artifacts contribute to # computationally enabled teaching and learning, and describing how they might be adopted by # others.
 # For learning modules, describe the learning objectives, content, instructional design, and # experience of use in teaching and learning situations.
 # Tell us the “story” of the project: how did it come to be?
-# Cite key references, including a link to the open archive of the sofware or the learning module.
+# Cite key references, including a link to the open archive of the software or the learning module.
 
 
 # # Citations
@@ -71,7 +71,7 @@ Managing the environment and natural resources is often faced with taking decisi
 
 Effective environmental management is often aided by using models. Such models face the challenges of limited data and the inherent simplifications of the complex natural world necessary to make them. This translates to uncertainties in critical decision-making predictions. Transparency is critical in reporting and mitigating these uncertainties, especially when significant risks accompany management outcomes. The significance of uncertainty quantification (UQ) and parameter estimation (PE) in environmental modeling for decision support is widely acknowledged [@Anderson2015]. UQ provides estimates of outcome uncertainty, while PE reduces this uncertainty through assimilating data, using a formal comparison of model outputs with observation data to inform model parameter values. 
 
-Implementing highly-parameterized UQ and PE in real-world modeling can be challenging due to both theoretical complexity and practical logistics. Limited project time and funding also often hinder their application. Open-source software such as ``PEST`` [@pest] and ``PEST++`` [@whitepestpp] provide tools for undertaking UQ and PE analyses. They have been employed to tackle a diverse range of real-world decision-support problems, ranging from risk-based well-head protection [@fienen2022risk], groundwater contaminant system remediation design [@FIENEN2024], evaluating the impact of groundwater use on river baseflow [@Foster2021],  managed aquiifer recharge schem design to protect against seawater intrusion [@Standen2022], optimizing vinyard irrigation [@KNOWLING2023108225], and many others. 
+Implementing highly-parameterized UQ and PE in real-world modeling can be challenging due to both theoretical complexity and practical logistics. Limited project time and funding also often hinder their application. Open-source software such as ``PEST`` [@pest] and ``PEST++`` [@whitepestpp] provide tools for undertaking UQ and PE analyses. They have been employed to tackle a diverse range of real-world decision-support problems, ranging from risk-based well-head protection [@fienen2022risk], groundwater contaminant system remediation design [@FIENEN2024], evaluating the impact of groundwater use on river baseflow [@Foster2021],  managed aquifer recharge scheme design to protect against seawater intrusion [@Standen2022], optimizing vineyard irrigation [@KNOWLING2023108225], and many others. 
 
 However, the steep learning curve associated with their use and the lack of user-friendly training materials have been a barrier to uptake. There is a growing demand within the environmental modelling community for transparent, reproducible, and accountable modeling processes, driven by the need for increased credibility and rigor in computational science and environmental simulation [@white11rapid; @fienen2016hess]. While some script-based tools enhance the reproducibility of forward model construction [@flopy], they often overlook UQ and PE analyses. In decision-support scenarios, these analyses are equally vital for robust model deployment as the forward model itself. 
 
@@ -84,8 +84,7 @@ The tutorial notebooks discussed herein provide a comprehensive, self-guided, an
 
[…]
