+++ paper/peer_review/responses_to_JOSE_reviews.md
@@ -1,0 +1,75 @@
+# Reply to https://github.com/openjournals/jose-reviews/issues/244
+
+Thank you @incsanchezro for your comprehensive review. Please see below replies and updates.
+
+Kind regards,
+-R
+
+> **Comments:**
+> 1. The authors have articulated the need for this educational framework, targeting the environmental modellers and decision-makers as specific audiences that could benefit from its implementation. The framework's usage and installation instructions are detailed. 
+
+Appreciated.
+
+> 2. A comprehensive workflow is recommended for the written and framework documentation. This addition should clearly outline the self-learning process, including learning objectives, module prerequisites, and potential implementations. It should make it easier for users to navigate the software autonomously or for educators to integrate it into their teaching methodologies, provided they possess the necessary expertise.
+
+We appreciate the suggestion and agree in principle. We had purposefully left it open ended, as as we tend to find that most users have unique issues that result in each taking their own path. 
+
+Nevertheless, we have added learning objectives and prerequisites in the "Contents and Instructional Design" section of the manuscript and to the README. Furthermore we have included an "example workflow" schematic outline. 
+
+We feel that the existing "workflow" outlined in the [README](https://github.com/gmdsi/GMDSI_notebooks/tree/jose-review#organization-of-the-tutorial-notebooks) provides sufficient guidance to the self-learning process. Furthermore, each of the Notebooks has an introductory section where the learning objectives are usually outlined, along with how these are related to real-world applications.
+
+> 3. The repository's current documentation needs specific guidelines for community involvement in the module's collaborative development. This oversight could be rectified by establishing clear pathways for third-party contributions, issue reporting, and support-seeking, fostering a more engaged and supportive community around the framework.
+
+We have included a "Contributing" section in the README. See commit https://github.com/gmdsi/GMDSI_notebooks/commit/bf5229a4884892a2b9af1e1f018ddde87e66e1d7
+
+> 4. Additionally, clarifying the project's funding sources and outlining strategies for its long-term sustainability would provide further transparency and confidence in the framework's continued evolution and support.
+
+Funding and support for the project is described in the Acknowledgements section of the README. See https://github.com/gmdsi/GMDSI_notebooks#acknowledgments
+
+
+# Reply to https://github.com/openjournals/jose-reviews/issues/245
+
+> As this repository has its origins in the classroom and was guided by teaching praxis, I > noticed its clarity regarding how it can be used as part of class teaching. My comments below > aim to enhance the pedagogical framework, making it clearer for both students and professors. > Furthermore, allowing its broader use of the repository. I hope you find them useful!
+> 
+> **Comments**
+> 
+> 1. While the educational modules provide valuable information, the learning objectives across > all sections need to be explicitly stated. It's recommended that the authors dedicate efforts > to delineating these objectives and clearly indicating the prerequisite knowledge necessary > for each module. This clarity will significantly enhance learners' ability to align their > expectations with the module's demands.
+
+We appreciate the suggestion and agree in principle. See response to https://github.com/openjournals/jose-reviews/issues/244
+
+> 2. The framework's current presentation on decision-making within the context of environmental > modelling requires clarity. The authors should further elaborate on their methodology and > precisely define what is encompassed by 'decision-making' in this specialized field. Such > clarification would give learners a more coherent understanding of the module's objectives > and applications.
+
+We have added clarifying text to the Statement of Need section of the paper.
+
+> 3. Furthermore, expanding the authors' pedagogical strategies regarding the support of learners > through the "how to use" and "how to think" processes could further enrich the learning > experience. I suggest amplification of these two elements in the written document and > specific activities into the modules that provide opportunities for the acquisition of skills > and knowledge.
+
[…]
