### 2024-05-10 #240 @codyalbertross
@kls2177

Thank you for your patience with my review.

_Self-Guided Decision Support Groundwater Modelling with Python_ is a valuable contribution and satisfies a need for concise and informed guidance on PEST and pyEMU. I agree with @incsanchezro regarding the organization of the workbook and the effort put forward by the authors to develop and improve GMDSI.

I found that the detail and theoretical background was exceptional and that the workbook followed a logical trajectory that is aligned with a typical workflow.

Overall, I believe that this work is suitable for publication and that the authors seem ready to take feedback from end-users to continuously improve their experience.
