### 2024-05-12 fa01d4f
add flowchart to paper
