### 2024-05-13 #240 @rhugman
Thank you @incsanchezro and @codyalbertross for your time and effort on the reviews.

We have addressed your comments and updated the repo and paper in the `jose-review` branch. I will leave direct responses to each of the specific comments.

All the best,
-R
