### 2024-05-13 #240 @kls2177
Hi @incsanchezro, thank you very much for your review. A technical request: can you please add your issues to the [groundwater resource repo](https://github.com/gmdsi/GMDSI_notebooks/), not the JOSE repo? Please see the [example that I provided above](https://github.com/openjournals/jose-reviews/issues/90#issuecomment-1601577547) - if you click on the links in this example, you will notice that the links go to the resource repo. Thanks!
