### 2024-05-15 75856ef
refactored resp surf ies to move nasty plotting func to script and import it. also a few typos. cleared notebooks
