### 2024-05-15 0df1013
updated flopy, pyemu, and env. trepidation
