+++ dependencies/flopy/mf6/modflow/mfutltvk.py
@@ -1,0 +1,287 @@
+# DO NOT MODIFY THIS FILE DIRECTLY.  THIS FILE MUST BE CREATED BY
+# mf6/utils/createpackages.py
+# FILE created on February 07, 2024 20:16:08 UTC
+from .. import mfpackage
+from ..data.mfdatautil import ListTemplateGenerator
+
+
+class ModflowUtltvk(mfpackage.MFPackage):
+    """
+    ModflowUtltvk defines a tvk package within a utl model.
+
+    Parameters
+    ----------
+    parent_package : MFPackage
+        Parent_package that this package is a part of. Package is automatically
+        added to parent_package when it is initialized.
+    loading_package : bool
+        Do not set this parameter. It is intended for debugging and internal
+        processing purposes only.
+    print_input : boolean
+        * print_input (boolean) keyword to indicate that information for each
+          change to the hydraulic conductivity in a cell will be written to the
+          model listing file.
+    timeseries : {varname:data} or timeseries data
+        * Contains data for the ts package. Data can be stored in a dictionary
+          containing data for the ts package with variable names as keys and
+          package data as values. Data just for the timeseries variable is also
+          acceptable. See ts package documentation for more information.
+    perioddata : [cellid, tvksetting]
+        * cellid ((integer, ...)) is the cell identifier, and depends on the
+          type of grid that is used for the simulation. For a structured grid
+          that uses the DIS input file, CELLID is the layer, row, and column.
+          For a grid that uses the DISV input file, CELLID is the layer and
+          CELL2D number. If the model uses the unstructured discretization
+          (DISU) input file, CELLID is the node number for the cell. This
+          argument is an index variable, which means that it should be treated
+          as zero-based when working with FloPy and Python. Flopy will
+          automatically subtract one when loading index variables and add one
+          when writing index variables.
+        * tvksetting (keystring) line of information that is parsed into a
+          property name keyword and values. Property name keywords that can be
+          used to start the TVKSETTING string include: K, K22, and K33.
+            k : [double]
+                * k (double) is the new value to be assigned as the cell's
+                  hydraulic conductivity from the start of the specified stress
+                  period, as per K in the NPF package. If the OPTIONS block
+                  includes a TS6 entry (see the "Time-Variable Input" section),
+                  values can be obtained from a time series by entering the
+                  time-series name in place of a numeric value.
+            k22 : [double]
+                * k22 (double) is the new value to be assigned as the cell's
+                  hydraulic conductivity of the second ellipsoid axis (or the
+                  ratio of K22/K if the K22OVERK NPF package option is
+                  specified) from the start of the specified stress period, as
+                  per K22 in the NPF package. For an unrotated case this is the
+                  hydraulic conductivity in the y direction. If the OPTIONS
+                  block includes a TS6 entry (see the "Time-Variable Input"
+                  section), values can be obtained from a time series by
+                  entering the time-series name in place of a numeric value.
+            k33 : [double]
+                * k33 (double) is the new value to be assigned as the cell's
+                  hydraulic conductivity of the third ellipsoid axis (or the
+                  ratio of K33/K if the K33OVERK NPF package option is
+                  specified) from the start of the specified stress period, as
+                  per K33 in the NPF package. For an unrotated case, this is
+                  the vertical hydraulic conductivity. If the OPTIONS block
+                  includes a TS6 entry (see the "Time-Variable Input" section),
+                  values can be obtained from a time series by entering the
+                  time-series name in place of a numeric value.
+    filename : String
+        File name for this package.
+    pname : String
+        Package name for this package.
[…]
