+++ dependencies/flopy/mf6/utils/testutils.py
@@ -1,0 +1,256 @@
+from ...utils import datautil
+
+
+def make_int_tuple(str_list):
+    int_list = []
+    for item in str_list:
+        int_list.append(int(item) - 1)
+    return tuple(int_list)
+
+
+def read_vertices(vert_file):
+    fd = open(vert_file, "r")
+    vertrecarray = []
+    for line in fd:
+        fd_spl = line.strip().split()
+        vertrecarray.append(
+            (int(fd_spl[0]) - 1, float(fd_spl[1]), float(fd_spl[2]))
+        )
+    fd.close()
+    return vertrecarray
+
+
+def read_cell2d(cell2d_file):
+    fd = open(cell2d_file, "r")
+    c2drecarray = []
+    for line in fd:
+        fd_spl = line.strip().split()
+        rec_array = [int(fd_spl[0]) - 1, float(fd_spl[1]), float(fd_spl[2])]
+        rec_array.append(int(fd_spl[3]))
+        for item in fd_spl[4:]:
+            rec_array.append(int(item) - 1)
+        c2drecarray.append(tuple(rec_array))
+    fd.close()
+    return c2drecarray
+
+
+def read_exchangedata(gwf_file, cellid_size=3, cellid_size_2=3):
+    exgrecarray = []
+    fd = open(gwf_file, "r")
+    for line in fd:
+        linesp = line.strip().split()
+        cellid_tot = cellid_size + cellid_size_2
+        exgrecarray.append(
+            (
+                make_int_tuple(linesp[0:cellid_size]),
+                make_int_tuple(linesp[cellid_size:cellid_tot]),
+                int(linesp[cellid_tot]),
+                float(linesp[cellid_tot + 1]),
+                float(linesp[cellid_tot + 2]),
+                float(linesp[cellid_tot + 3]),
+                float(linesp[cellid_tot + 4]),
+            )
+        )
+    return exgrecarray
+
+
+def read_gncrecarray(gnc_file, cellid_size=3, cellid_size_2=3):
+    gncrecarray = []
+    fd = open(gnc_file, "r")
+    for line in fd:
+        linesp = line.strip().split()
+        gncrecarray.append(
+            (
+                make_int_tuple(linesp[0:cellid_size]),
+                make_int_tuple(
+                    linesp[cellid_size : cellid_size + cellid_size_2]
+                ),
+                make_int_tuple(
+                    linesp[
+                        cellid_size + cellid_size_2 : cellid_size * 2
+                        + cellid_size_2
+                    ]
+                ),
+                float(linesp[cellid_size * 2 + cellid_size_2]),
+            )
+        )
+    return gncrecarray
+
+
+def read_chdrecarray(chd_file, cellid_size=3):
+    fd = open(chd_file, "r")
+    chdrecarray = []
+    for line in fd:
+        fd_spl = line.strip().split()
+        chdrecarray.append(
+            (make_int_tuple(fd_spl[0:cellid_size]), float(fd_spl[cellid_size]))
+        )
+    fd.close()
+    return chdrecarray
+
+
+def read_ghbrecarray(chd_file, cellid_size=3):
+    fd = open(chd_file, "r")
+    ghbrecarray = []
+    for line in fd:
+        fd_spl = line.strip().split()
+        ghbrecarray.append(
+            (
+                make_int_tuple(fd_spl[0:cellid_size]),
+                float(fd_spl[cellid_size]),
+                float(fd_spl[cellid_size + 1]),
+            )
+        )
+    fd.close()
+    return ghbrecarray
+
+
+def read_obs(obs_file, cellid_size=3):
+    fd = open(obs_file, "r")
+    obsrecarray = []
+    for line in fd:
[…]

[…]
