### 2024-05-16 b6a0713
updated successfully env and deps. ran all tests. Made a small mod to flopy to suppress a warning brought on by paertial pandas implementation attempt in the flopy code
