+++ dependencies/pyemu/README.md
@@ -1,0 +1,136 @@
+pyEMU
+=====
+
+python modules for model-independent FOSM (first-order, second-moment) (a.k.a linear-based, a.k.a. Bayes linear) 
+uncertainty analyses and data-worth analyses, non-linear uncertainty analyses and interfacing with PEST and PEST++.  
+pyEMU also has a pure python (pandas and numpy) implementation of ordinary kriging for geostatistical interpolation and 
+support for generating high-dimensional PEST(++) model interfaces, including support for (very) high-dimensional 
+ensemble generation and handling   
+
+Main branch:
+[![pyemu continuous integration](https://github.com/pypest/pyemu/actions/workflows/ci.yml/badge.svg?branch=main)](https://github.com/pypest/pyemu/actions/workflows/ci.yml)
+[![Coverage Status](https://coveralls.io/repos/github/pypest/pyemu/badge.svg?branch=main)](https://coveralls.io/github/pypest/pyemu?branch=main)
+
+Develop branch:
+[![pyemu continuous integration](https://github.com/pypest/pyemu/actions/workflows/ci.yml/badge.svg?branch=develop)](https://github.com/pypest/pyemu/actions/workflows/ci.yml)
+[![Coverage Status](https://coveralls.io/repos/github/pypest/pyemu/badge.svg?branch=develop)](https://coveralls.io/github/pypest/pyemu?branch=develop)
+
+Documentation
+=============
+
+Complete user's guide:
+
+[https://pyemu.readthedocs.io/en/latest/](https://pyemu.readthedocs.io/en/latest/)
+
+The pyEMU documentation is being treated as a first-class citizen!  Also see the example notebooks in the repo.
+
+What is pyEMU?
+================
+
+pyEMU is a set of python modules for model-independent, user-friendly, computer model uncertainty analysis.  pyEMU is tightly coupled to the open-source suite PEST (Doherty 2010a and 2010b, and Doherty and other, 2010) and PEST++ (Welter and others, 2015, Welter and other, 2012), which are tools for model-independent parameter estimation.  However, pyEMU can be used with generic array objects, such as numpy ndarrays.
+
+Several equations are implemented, including Schur's complement for conditional uncertainty propagation (a.k.a. Bayes Linear estimation) (the foundation of the PREDUNC suite from PEST) and error variance analysis (the foundation of the PREDVAR suite of PEST).  pyEMU has easy-to-use routines for parameter and data worth analyses, which estimate how increased parameter knowledge and/or additional data effect forecast uncertainty in linear, Bayesian framework.  Support is also provided for high-dimensional Monte Carlo analyses via `ObservationEnsemble` and `ParameterEnsemble` class, including the null-space monte carlo approach of Tonkin and Doherty (2009); these ensemble classes also play nicely with PESTPP-IES.
+
+pyEMU also includes lots of functionality for dealing with PEST(++) datasets, such as:
+* manipulation of PEST control files, including the use of pandas for sophisticated editing of the parameter data and observation data sections
+* creation of PEST control files from instruction and template files
+* going between site sample files and pandas dataframes - really cool for observation processing
+* easy-to-use observation (re)weighting via residuals or user-defined functions
+* handling Jacobian and covariance matrices, including functionality to go between binary and ASCII matrices, reading and writing PEST uncertainty files.  Covariance matrices can be instantiated from relevant control file sections, such as parameter bounds or observation weights.  The base Matrix class overloads most common linear algebra operators so that operations are automatically aligned by row and column name.  Builtin SVD is also included in all Matrix instances.
[…]
