### 2024-05-17 a62804b
Update README.md
quick change to recommend miniforge rather than anaconda
