+++ README.md
@@ -72,7 +72,7 @@ You can do this in one of two ways.
 **Install Python and dependencies:**
  - If you have already installed Python using Anaconda, you can skip this step. If not, install [miniforge](https://github.com/conda-forge/miniforge?tab=readme-ov-file#download).  
  - If you are using __Windows__: go to the start menu and open "Anaconda prompt". An anaconda command line window will open. On __Linux__ or __MacOS__, just use the standard terminal. Navigate to the course repo folder on your machine. You can accomplish this by typing "cd *your folder path*" and pressing < enter >. Replace *your folder path* with the path to the course material folder on your computer.
- - Next, type `mamba env create -f environment.yml` (or `conda env create -f environment.yml` if `mamba` doesn't work for you. This will create an anaconda environment called "gmdsitut" and install the python dependencies required for this course. It may take a while. Should you wish, you can inspect the *environment.yml* file in the repo folder to see what dependecies are being installed.
+ - Next, type `mamba env create -f environment.yml` (or `conda env create -f environment.yml` if `mamba` doesn't work for you). This will create an anaconda environment called "gmdsitut" and install the python dependencies required for this course. It may take a while. Should you wish, you can inspect the *environment.yml* file in the repo folder to see what dependecies are being installed.
 
 **Start jupyter notebook**
 You will need to do this step any time you wish to open one of the course notebooks.
