+++ paper/paper.bib
@@ -128,6 +128,7 @@ doi = "http://dx.doi.org/10.1016/B978-0-08-091638-5.00010-9"
   author={White, Jeremy and Fienen, Michael and Moore, Catherine and Guthke, Anneli},
   journal={Frontiers in Earth Science},
   volume={11},
+  year="2023",
   pages={1260581},
   publisher={Frontiers},
   doi = {10.3389/978-2-8325-3581-3}
@@ -247,9 +248,23 @@ doi = "http://dx.doi.org/10.1016/B978-0-08-091638-5.00010-9"
   pages = {108225},
   year = {2023},
   issn = {0378-3774},
-  doi = {https://doi.org/10.1016/j.agwat.2023.108225},
+  doi = {10.1016/j.agwat.2023.108225},
   url = {https://www.sciencedirect.com/science/article/pii/S0378377423000902},
   author = {Matthew J. Knowling and Rob R. Walker and Anne Pellegrino and Everard J. Edwards and Seth Westra and Cassandra Collins and Bertram Ostendorf and Bree Bennett},
   keywords = {Decision agriculture, Water production function, Crop model, Viticulture, Irrigation},
   abstract = {The potential of digital agriculture to support on-farm decision making is predicated on the assumption that ‘cause-and-effect’ relationships can be encoded in a mathematical form. One particularly important application area is irrigation decision making, which is informed by the relationship between applied water and end-of-season crop yield (‘water production relations’). Yet this relationship is often partial, owing to its many determining factors, especially for woody perennial crops such as grapevines. Process-based models are a way in which to represent these relationships in a manner that is both interpretable and generalizable. Here we conduct numerical experiments using a process-based crop model to evaluate water production relations for grapevines and how these relations are influenced by genetic and environmental factors as well as irrigation timing decisions. A real-world case study representing a Shiraz vineyard in South Australia is considered. Results show a largely linear relation between total irrigation applied and yield across all numerical experiments, notwithstanding significant uncertainty due to genetic and environmental factors. However, when considering water production relations in relative terms (e.g., change in tonnes per megalitre), the influence of these factors between seasons is reduced, allowing for more robust insights. Exploration of water productivity as a function of phenological stage shows that the average production sensitivity is greatest during veraison (3.5 tonnes per megalitre) and least between bud burst and flowering (2.3 tonnes per megalitre), despite considerable overlap in productivity range between stages. By putting meaningful bounds on water production relations through process-based modeling, growers and their advisors can achieve improved farm outcomes by better informed water application decisions.}
   }
+  
+@article{freyberg,
+author = {Freyberg, David L.},
+title = {AN EXERCISE IN GROUND-WATER MODEL CALIBRATION AND PREDICTION},
+journal = {Groundwater},
+volume = {26},
+number = {3},
+pages = {350-360},
+doi = {10.1111/j.1745-6584.1988.tb00399.x},
+url = {https://ngwa.onlinelibrary.wiley.com/doi/abs/10.1111/j.1745-6584.1988.tb00399.x},
+eprint = {https://ngwa.onlinelibrary.wiley.com/doi/pdf/10.1111/j.1745-6584.1988.tb00399.x},
[…]
