+++ paper/paper.md
@@ -75,7 +75,7 @@ Implementing highly parameterized UQ and PE in real-world modelling can be chall
 
 However, the steep learning curve associated with their use and the lack of user-friendly training materials have been a barrier to uptake. There is a growing demand within the environmental modelling community for transparent, reproducible, and accountable modelling processes, driven by the need for increased credibility and rigor in computational science and environmental simulation [@white11rapid; @fienen2016hess]. While some script-based tools enhance the reproducibility of forward model construction [@flopy], they often overlook UQ and PE analyses. In decision-support scenarios, these analyses are equally vital for robust model deployment as the forward model itself. 
 
-The uptake of Python for environmental modelling has increased in recent years, due to its open-source nature, user-friendly syntax, and extensive scientific libraries. Python-based tools have been developed to facilitate UQ and PE analyses, such as ``pyemu`` [@White_A_Python_framework_2016; @white2021towards]. ``pyemu`` is a Python package that provides a framework for implementing UQ and PE analyses with ``PEST`` and ``PEST++``. It offers a range of capabilities, including parameter estimation, uncertainty analysis, and management optimization. Although initially designed for groundwater modelling, ``pyemu``'s methodologies are versatile and can be applied to diverse numerical environmental models, as long as they use text files for input and generate machine-readable outputs that can be extracted without manual intervention.
+The uptake of Python for environmental modelling has increased in recent years, due to its open-source nature, user-friendly syntax, and extensive scientific libraries. Python-based tools have been developed to facilitate UQ and PE analyses, such as ``pyemu`` [@White_A_python_framework_2016; @white2021towards]. ``pyemu`` is a Python package that provides a framework for implementing UQ and PE analyses with ``PEST`` and ``PEST++``. It offers a range of capabilities, including parameter estimation, uncertainty analysis, and management optimization. Although initially designed for groundwater modelling, ``pyemu``'s methodologies are versatile and can be applied to diverse numerical environmental models, as long as they use text files for input and generate machine-readable outputs that can be extracted without manual intervention.
 
 The tutorial notebooks discussed herein provide a comprehensive, self-guided, and open-source resource for learning decision-support modelling workflows with Python. They are designed to be accessible to a broad audience, including students, researchers, and practitioners who aim to undertake applied environmental decision-support modelling. 
 
@@ -92,7 +92,7 @@ Decision-support modelling often demands capabilities that surpass what current 
 
 Anecdotally, we have seen that more modelers are turning to Python packages like ``FloPy`` [@flopy] and ``pyemu`` [@whitepyemu] for model and ``PEST++`` setup. Unfortunately, the adoption of this approach is hindered by a steep learning curve primarily due to the scarcity of user-friendly training materials. The [GMDSI tutorial notebooks](https://github.com/gmdsi/GMDSI_notebooks/tree/main) aim to address this gap by providing a comprehensive, self-guided, and open-source resource for learning decision-support modelling workflows with Python.
 
[…]
