### 2024-06-03 #240 @rhugman
Hi @kls2177,

We have updated the repo in the [ `jose-review`](https://github.com/gmdsi/GMDSI_notebooks/tree/jose-review) branch addressing @incsanchezro and @codyalbertross comments. We are not clear on what the next steps are, can you please let us know what we need to do next?
