### 2024-06-13 07c005c
added course registration to readme for upcoming class
