### 2024-07-24 5b73ca8
update env.yl for numpy <2
