+++ environment.yml
@@ -5,7 +5,7 @@ dependencies:
   - python=3.11
   - pip
   - pandas>=2.0
-  - numpy>=1.25 
+  - numpy<2 
   - matplotlib 
   - openpyxl
   - xlrd
