### 2024-07-29 4b5437f
Removing next course offering from README now that course is in progress
