+++ dependencies/pyemu/pyemu/eds.py
@@ -11,7 +11,7 @@ from pyemu.en import ObservationEnsemble
 from pyemu.mat.mat_handler import Matrix, Jco, Cov
 from pyemu.pst.pst_handler import Pst
 from pyemu.utils.os_utils import _istextfile,run
-from pyemu.utils.helpers import normal_score_transform,randrealgen_optimized,inverse_normal_score_transform
+from pyemu.utils.helpers import normal_score_transform,randrealgen_optimized
 from .logger import Logger
 
 
@@ -496,7 +496,9 @@ class EnDS(object):
         return mean_dfs,dfstd,dfper
 
 
-    def prep_for_dsi(self,sim_ensemble=None,t_d="dsi_template",apply_normal_score_transform=False,truncated_svd=False):
+    def prep_for_dsi(self,sim_ensemble=None,t_d="dsi_template",
+                     apply_normal_score_transform=False,
+                     use_ztz=False,energy=1.0,nst_extrap=None):
         """Setup a new PEST interface for the data-space inversion process.
         If the observation data in the Pst object has a "obstransform" column, then observations for which "log" is specified will be subject to log-transformation. 
         If the `apply_normal_score_transform` flag is set to `True`, then the observations and predictions will be subject to a normal score transform.
@@ -510,6 +512,7 @@ class EnDS(object):
                 and predictions.  Default is `False`
             truncated_svd (`bool`): flag to use a truncated SVD for the pseudo-inverse of the deviations matrix.
                 Default is `False`
+            nst_extrap (None or 'str'): flag to select extrapolation type used during normal-score transformation. Can be None, "linear" or "quadratic". If None, normal-score back-transformation is truncated to the range of the distribution of the provided observation ensemble.
 
         Example::
 
@@ -530,7 +533,7 @@ class EnDS(object):
             shutil.rmtree(t_d)
         os.makedirs(t_d)
 
-        self.logger.log("getting deviations")
+        
         nz_names = self.pst.nnz_obs_names
         snz_names = set(nz_names)
         z_names = [n for n in self.pst.obs_names if n not in snz_names]
@@ -538,34 +541,49 @@ class EnDS(object):
         names.extend(nz_names)
         names.sort()
 
+        # make sure names are sorted
+        sim_ensemble = sim_ensemble.loc[:,names]
+        
+        self.logger.log("applying transformations")
         # implement log-transform/offset and normal score transform
         transf_names = nz_names.copy()
         transf_names.extend(self.predictions)
         
         if "obstransform" in self.pst.observation_data.columns:
-            obs = self.pst.observation_data.loc[transf_names].copy()
+            obs = self.pst.observation_data.copy()
+            #make sure names are ordered
+            obs = obs.loc[names,:]
+            #TODO: deal with "scale" and user-specified "offset"
             obs["offset"] = 0.0 #TODO: more elegant? incase all 'none' are passed...
             obsnmes = obs.loc[obs.obstransform=='log'].obsnme.values
             if len(obsnmes) > 0:
                 for name in obsnmes:
                     #TODO: make more efficient
                     self.logger.log("applying obs log-transform to:"+name)
-                    values = sim_ensemble.loc[:,name].values
+                    values = sim_ensemble.loc[:,name].astype(float).values
                     offset = abs(min(values))+1.0 #arbitrary; enforce positive values
                     values+=offset
[…]
