+++ dependencies/pyemu/pyemu/eds.py
@@ -497,8 +497,8 @@ class EnDS(object):
 
 
     def prep_for_dsi(self,sim_ensemble=None,t_d="dsi_template",
-                     apply_normal_score_transform=False,
-                     use_ztz=False,energy=1.0,nst_extrap=None):
+                     apply_normal_score_transform=False,nst_extrap=None,
+                     use_ztz=False,energy=1.0):
         """Setup a new PEST interface for the data-space inversion process.
         If the observation data in the Pst object has a "obstransform" column, then observations for which "log" is specified will be subject to log-transformation. 
         If the `apply_normal_score_transform` flag is set to `True`, then the observations and predictions will be subject to a normal score transform.
@@ -510,9 +510,11 @@ class EnDS(object):
             t_d (`str`): template directory to setup the DSI model + pest files in.  Default is `dsi_template`
             apply_normal_score_transform (`bool`): flag to apply a normal score transform to the observations
                 and predictions.  Default is `False`
-            truncated_svd (`bool`): flag to use a truncated SVD for the pseudo-inverse of the deviations matrix.
+            nst_extrap (`str`): flag to apply extrapolation to the normal score transform. Can be None, 'linear' or 'quadratic'. Default is None. 
+            use_ztz (`bool`): flag to use the condensed ZtZ matrix for SVD. The ZtZ matrix has dimensions nreal*nreal, instead of the nreal*nobs dimensions of Z. 
+                This makes the SVD computation faster and more memory efficient when nobs >> nreal. Default is `False`
                 Default is `False`
-            nst_extrap (None or 'str'): flag to select extrapolation type used during normal-score transformation. Can be None, "linear" or "quadratic". If None, normal-score back-transformation is truncated to the range of the distribution of the provided observation ensemble.
+            energy (`float`): energy threshold for truncating the sqrt(C) matrix.  Default is `1.0` which applys no truncation. 
 
         Example::
 
@@ -613,18 +615,16 @@ class EnDS(object):
         if use_ztz:
             self.logger.log("using ztz approach...")
             pmat, s = compute_using_ztz(sim_ensemble)
-            nobs = sim_ensemble.shape[1]
-            nreal = sim_ensemble.shape[0]
             self.logger.log("using ztz approach...")
-
-            #self.logger.log("applying truncation...")
-            apply_energy_based_truncation(energy, s, pmat,nobs,nreal)
-            #self.logger.log("applying truncation...")
         else:
             self.logger.log("using z approach...")
             pmat, s = compute_using_z(sim_ensemble)
             self.logger.log("using z approach...")
         self.logger.log("computing projection matrix")
+
+        self.logger.log("applying truncation...")
+        apply_energy_based_truncation(energy,s,pmat)
+        self.logger.log("applying truncation...")
 
         self.logger.log("creating tpl files")
         dsi_in_file = os.path.join(t_d, "dsi_pars.csv")
@@ -743,9 +743,16 @@ class EnDS(object):
         par = pst.parameter_data
         dsi_pars = par.loc[par.parnme.str.startswith("dsi_par"),"parnme"]
         par.loc[dsi_pars,"parval1"] = 0
-        par.loc[dsi_pars,"parubnd"] = 2.5
-        par.loc[dsi_pars,"parlbnd"] = -2.5
[…]
