### 2024-08-19 4e46f8a
more updates to pyemu dsi deps
