### 2024-09-06 #240 @kls2177
> Hi @kls2177,
>
> We have updated the repo in the [ `jose-review`](https://github.com/gmdsi/GMDSI_notebooks/tree/jose-review) branch addressing @incsanchezro and @codyalbertross comments. We are not clear on what the next steps are, can you please let us know what we need to do next?

Hi @rhugman, to speed the process along, can you please provide links to your responses to the comments by reviews in the chat? Thanks!
