### 2024-09-06 #240 @rhugman
Hi @kls2177, thanks for expediting this.

We responded to comments by @incsanchezro directly in the issues they raised (linked below):

> 1. > > [JOSE Review - comments on "Self-Guided Decision Support Groundwater Modelling with Python " [Documentation section] #244](https://github.com/openjournals/jose-reviews/issues/244)
> 2. > > [JOSE Review - comments on "Self-Guided Decision Support Groundwater Modelling with Python " [Pedagogy / Instructional design ] #245](https://github.com/openjournals/jose-reviews/issues/245)
> 3. > > [JOSE Review - comments on "Self-Guided Decision Support Groundwater Modelling with Python " [JOSE paper] #246](https://github.com/openjournals/jose-reviews/issues/246)
