### 2024-09-13 #240 @kls2177
Hi @rhugman,

   While we wait for any final comments from @incsanchezro, I have a few comments of my own.

1. The JOSE paper is over the [suggested word limit of 1000 words](https://openjournals.readthedocs.io/en/jose/submitting.html#:~:text=Expected%20length%20is%20around%201000%20words%20max.). I encourage the authors to revise and shorten this paper. Some of my suggestions include moving the Resources section and the Contents section to the tutorial website.

Also, I recommend a quick grammatical/spell check of the paper. There are a couple of minor issues.

2. The learning outcomes for each section in the tutorials need to be more explicit and clearly stated at the outset of each section. This is sometimes the case, but is not consistent throughout. A stylistic option could also be to place the learning outcomes within a ["special content blocks"](https://jupyterbook.org/en/stable/content/content-blocks.html) so that they are clearly marked for students.

3. For each Chapter, I recommend removing the table of contents landing page (as this is visible in the side-bar) and replace with a narrative landing page that provides the big picture context and lays out the connections between the sections within the chapter.

4. Finally, it would be helpful for students to hyperlink within your tutorial to earlier chapters/sections of the tutorial to help students make the connections between topics.
