### 2024-09-21 5f86616
moved resources and contents to readme.md
