+++ paper/paper.md
@@ -96,88 +96,6 @@ The roots of the materials making up the tutorial notebooks were from a traditio
 
 After three iterations of teaching the in-person class, the instructors concluded that the materials and approach were valuable, but came to question the level of retention by students in a 40-hour intensive setting. It is well-documented that without repetition and rapid adoption of new techniques, they can fade quickly from memory [@glaveski2019companies]. As a result, the authors, with support from the GMDSI, endeavored to build on the positive aspects of using Jupyter Notebooks and explore alternative teaching environments instead of week-long classes. The first major change was to add sufficient narration and explanation to the notebooks to improve possibilities for self-study. The initial design through in-person instruction was to have the notebooks serve as illustrations to assist in a narrative discussion, so bolstering of the explanatory text was necessary to help them stand alone. The next change was to refactor the organization from a strictly linear progression to the current three-part organization discussed below. This led to a hybrid model of self-study punctuated by discussion and background lectures online. 
 
-# Resources
-
-A webinar hosted by GMDSI introducing the tutorial notebooks can be viewed [here](https://vimeo.com/856752189). During the webinar the authors provided an overview of the notebooks, as well as a demonstration of how to use them and introduced an [online self-guided course](https://gmdsi.org/blog/guided-self-study-course/).
-
-The [GMDSI](https://gmdsi.org) web-page also hosts an extensive range of resources and educational material on decision support modelling. These include numerous instructional video lectures, webinar recordings, non-programmatic workflow tutorials, as well as worked example reports describing real-world applications. 
-
-Software from the ``PEST`` suite can be downloaded from John Doherty's web page [here](https://www.pesthomepage.org/). The [user manual](https://www.pesthomepage.org/Downloads/PEST%20Manuals/PEST%20Manual.pdf) contains much useful information. The [PEST Book](https://pesthomepage.org/pest-book) is also a great resource for learning about the theory underpinning use of the software.
-
-Software from the ``PEST++`` suite can be accessed from GitHub [repository](https://github.com/usgs/pestpp/tree/master). The [user manual](https://github.com/usgs/pestpp/blob/develop/documentation/pestpp_users_manual.md) contains much useful information, as well as theoretical background to the software. Further theoretical background is available in [@whitepestpp].
-
-``pyemu`` can be accessed from the GitHub [repository](https://github.com/pypest/pyemu/tree/master). The repository contains several example Jupyter Notebooks. The tutorial notebooks discussed herein provide a more exhaustive and structured learning experience.
-
-
-# Contents and Instructional Design
-The tutorial notebooks are structured into three main parts:
-
-## Part 0: Introductory Background
-
-Part 0 serves as the foundation, providing essential background material. Learners are encouraged to reference notebooks in Part 0 to polish their understanding of concepts they encounter in Parts 1 and 2. Part 0 is not intended to be a comprehensive resource for all background material, but rather to establish a solid understanding of the basics. The explanations of mathematical concepts are intended to be accessible through visualization and descriptions related to everyday concepts and modelling concepts. 
-
-Each notebook in Part 0 is standalone and covers a unique topic. These include:
-
-  - Introduction to a synthetic model known as the "Freyberg" model [@freyberg]. This model is used as a consistent example throughout the tutorial exercises, allowing learners to apply concepts in a practical context.
-  - An introduction to the `pyemu` Python package that is used to complement and interface with PEST/PEST++.
-  - Explanation of fundamental mathematical concepts that are relevant and will be encountered throughout the tutorial notebooks.
-
[…]
