+++ README.md
@@ -78,13 +78,6 @@ PEST++MOU:
  
 IMPORTANT NOTE: When running some of these notebooks, PEST++ undertakes parallel model runs for computational efficiency. These runs are coordinated using internet communication protocols (e.g. TCP/IP). As a result, it is common (but not universal) that the first time a set of parallel runs is launched, a Windows Defender (or similar popup on other operating systems) box appears warning the user. This box can simply be closed by clicking "cancel" and then the model runs should proceed without issue.
 
-## Example Tutorial Notebook Learning Flow
-While there is an intentional organization to the notebooks, described above, they can also serve as standalone references, or a subset can be curated to achieve a specific learning outcome. An example of a curated learning flow for a common decision support modeling application is shown in the flowchart below. 
-
-<p align="center">
-  <img src="flowchart.png" alt="flowchart">
-</p>
-
 ## Installation Instructions
 
 **Download the course repository:**
@@ -105,8 +98,97 @@ To start up the jupyter notebook:
 - Mac/Linux: open a termainal and type `conda activate gmdsitut`
 - Then navigate to folder where you downloaded the course materials repo and type `jupyter notebook`
 A jupyter notebook instance should start within the course repo flder. Using the browser, you can now navigate to the "notebooks" folder and open one.
+
+
+
+# Contents and Instructional Design
+The tutorial notebooks are structured into three main parts:
+
+## Part 0: Introductory Background
+
+Part 0 serves as the foundation, providing essential background material. Learners are encouraged to reference notebooks in Part 0 to polish their understanding of concepts they encounter in Parts 1 and 2. Part 0 is not intended to be a comprehensive resource for all background material, but rather to establish a solid understanding of the basics. The explanations of mathematical concepts are intended to be accessible through visualization and descriptions related to everyday concepts and modelling concepts. 
+
+Each notebook in Part 0 is standalone and covers a unique topic. These include:
+
+  - Introduction to a synthetic model known as the "Freyberg" model [@freyberg]. This model is used as a consistent example throughout the tutorial exercises, allowing learners to apply concepts in a practical context.
+  - An introduction to the `pyemu` Python package that is used to complement and interface with PEST/PEST++.
+  - Explanation of fundamental mathematical concepts that are relevant and will be encountered throughout the tutorial notebooks.
+
+Pre-requisites for Part 0 include a basic understanding of Python, Jupyter Notebooks, and MODFLOW 6 [@Langevin_MODFLOW_6_Modular_2022]. Familiarity with git is a bonus but not fundamental.
+
+After completing this module, learners should be able to:
+
+  - Describe the structure and components of the synthetic Freyberg model.
+  - Apply the `pyemu` Python package to interface with PEST/PEST++.
+  - Discuss the mathematical concepts that underpin the PEST/PEST++ software.
+
+## Part 1: Introduction to ``PEST`` and the Gauss-Levenberg Marquardt Approach
+
+Part 1 focuses on the Gauss-Levenberg Marquardt (GLM) approach to parameter estimation and associated uncertainty analysis in a groundwater modelling context. This was the foundation of the ``PEST`` software for multiple decades and the theory continues to resonate through newer techniques.
+
+Part 1 is designed to be accessible without strict sequential dependencies. Learners have the flexibility to explore its contents in any order that suits their preferences or needs. These include:
+
+  - Introduction to concepts such as non-uniqueness, identifiability, and equifinality.
+  - Introduction to the ``PEST`` control file and the PEST/PEST++ interface.
+  - Exploring the challenges of parameterization schemes on predictive ability, as well as how to mitigate them.
+  - Introducing first-order second-moment (FOSM) and prior Monte Carlo uncertainty analysis approaches.
+
[…]
