+++ paper/paper.md
@@ -62,14 +62,14 @@ bibliography: paper.bib
 
 # Summary
 
-The [GMDSI tutorial notebooks repository](https://github.com/gmdsi/GMDSI_notebooks/tree/main) provides learners with a comprehensive set of tutorials for self-guided training on decision-support groundwater modelling using Python-based tools. Although targeted at groundwater modelling, they are based around model-agnostic tools and readily transferable to other environmental modelling workflows. The tutorials are divided into three parts. The first covers fundamental theoretical concepts. These are intended as background reading for reference on an as-needed basis. Tutorials in the second part introduce learners to some of the core concepts parameter estimation in a groundwater modelling context, as well as providing a gentle introduction to the ``PEST``, ``PEST++`` and ``pyemu`` software. Lastly, the third part demonstrates how to implement highly parameterized applied decision-support modelling workflows. The tutorials aim to provide examples of both “how to use” the software as well as “how to think” about using the software. 
+The [GMDSI tutorial notebooks repository](https://github.com/gmdsi/GMDSI_notebooks/tree/main) provides learners with a comprehensive set of tutorials for self-guided training on decision-support groundwater modelling using Python-based tools. Although targeted at groundwater modelling, they are based around model-agnostic tools and readily transferable to other environmental modelling workflows. The tutorials are divided into three parts. The first covers fundamental theoretical concepts. These are intended as background reading for reference on an as-needed basis. Tutorials in the second part introduce learners to some of the core concepts of parameter estimation in a groundwater modelling context, as well as provide a gentle introduction to the ``PEST``, ``PEST++`` and ``pyemu`` software. Lastly, the third part demonstrates how to implement highly parameterized applied decision-support modelling workflows. The tutorials aim to provide examples of both “how to use” the software as well as “how to think” about using the software. 
 A key advantage to using notebooks in this context is that the workflows described run the same code as practitioners would run on a large-scale real-world application. Using a small synthetic model facilitates rapid progression through the workflow.
 
 # Statement of Need
 
-Managing the environment and natural resources is often faced with making decisions on actions that trade off between economic, social, and environmental outcomes. In a groundwater management context, they may include decisions on such things as how much to pump from a well without impacting a nearby stream or how to design a well-field to maximize yield while minimizing the risk of saltwater intrusion. These decisions are often complex and rely on the use of numerical models to simulate the natural system and evaluate the outcomes of different management actions [@kelly2013]. 
+Managing the environment and natural resources is often faced with making decisions on actions that trade-off between economic, social, and environmental outcomes. In a groundwater management context, they may include decisions on such things as how much to pump from a well without impacting a nearby stream or how to design a well-field to maximize yield while minimizing the risk of saltwater intrusion. These decisions are often complex and rely on the use of numerical models to simulate the natural system and evaluate the outcomes of different management actions [@kelly2013]. 
 
-Effective environmental management is often aided by using models. Such models face the challenges of limited data and the inherent simplifications of the complex natural world necessary to make them. This translates to uncertainties in critical decision-making predictions. Transparency is critical in reporting and mitigating these uncertainties, especially when significant risks accompany management outcomes. The significance of uncertainty quantification (UQ) and parameter estimation (PE) in environmental modelling for decision support is widely acknowledged [@Anderson2015]. UQ provides estimates of outcome uncertainty, while PE reduces this uncertainty through assimilating data, using a formal comparison of model outputs with observation data to inform model parameter values. 
[…]
