### 2024-09-23 1ff2cbc
moved learning outcomes to special content block
