### 2024-09-23 #240 @rhugman
Thanks @kls2177,

> 1. The JOSE paper is over the [suggested word limit of 1000 words](https://openjournals.readthedocs.io/en/jose/submitting.html#:~:text=Expected%20length%20is%20around%201000%20words%20max.). I encourage the authors to revise and shorten this paper. Some of my suggestions include moving the Resources section and the Contents section to the tutorial website.
> Also, I recommend a quick grammatical/spell check of the paper. There are a couple of minor issues.

As requested,:
- [Resources and Contents section moved to the github landing page.](https://github.com/gmdsi/GMDSI_notebooks/commit/5f86616309fbe55ce38799ec626c28b81b1bf52a)
- [ran through spell checker](https://github.com/gmdsi/GMDSI_notebooks/commit/2eb986b25a88258e12540b3f511f2df4b31f9175)
- [significantly shortened text. Still slightly over 1k words. ](https://github.com/gmdsi/GMDSI_notebooks/commit/8c2f96fb4fe134809f5c50cd3a7f5c85eaa2310d)

> 2. The learning outcomes for each section in the tutorials need to be more explicit and clearly stated at the outset of each section. This is sometimes the case, but is not consistent throughout. A stylistic option could also be to place the learning outcomes within a ["special content blocks"](https://jupyterbook.org/en/stable/content/content-blocks.html) so that they are clearly marked for students.

[Placed learning outcomes in special content block](https://github.com/gmdsi/GMDSI_notebooks/commit/1ff2cbc8d08bf45cf9a63295d47d164808509702)

> 3. For each Chapter, I recommend removing the table of contents landing page (as this is visible in the side-bar) and replace with a narrative landing page that provides the big picture context and lays out the connections between the sections within the chapter.

I am not quite clear on what 3. and 4. refer to. Were these in regard to the GitHub Page? That page was very much out of date and has since been taken down. It was never part of the submitted paper. The only "website" that we maintain as part of this project is the git-hub repo.

> 4. Finally, it would be helpful for students to hyperlink within your tutorial to earlier chapters/sections of the tutorial to help students make the connections between topics.

Within the notebooks, where relevant we provide links to other specific notebooks.

Please let me know if that addresses your concerns.

Regards,
R
