### 2024-09-25 #240 @kls2177
Thanks @rhugman. The [link to the notebooks](https://gmdsi.github.io/GMDSI_notebooks/) is not working for me today.
