### 2024-09-25 #240 @rhugman
Hi @kls2177, as I mentioned in the previous post, that website was out of date and not intended as part of this submission. (Thank you for bringing to our attention that it was still live.) Apologies for any confusion this has caused - we missed that that was still there.

I have removed reference to it from the repo's landing page. As is described in the README.md, the tutorials are intended to be cloned/downloaded and run locally. The compute requirements are too large to make it feasible to host up-to-date versions of the notebooks on GitHub.
