### 2024-09-25 #240 @kls2177
Thanks for the clarification @rhugman. I had been looking at the website.

Thank you for revising the paper. I apologize for this, but I think it would be worth putting the "Content" section back in the paper. It will make the paper longer, but something is lost now that it is removed.

Once this is in place, we can move ahead with publication. Thank you for your patience!!
