### 2024-09-26 70b2587
put contents back in paper
