+++ paper/paper.md
@@ -88,6 +88,79 @@ Decision-support modelling often demands capabilities that surpass what current 
 
 The roots of the materials making up the tutorial notebooks were from a traditional, week-long classroom course curriculum developed for internal training at the US Geological Survey (USGS) by a subset of the authors of this paper. After three iterations of teaching the in-person class, the instructors concluded that the materials and approach were valuable, but came to question the level of retention by students in a 40-hour intensive setting. As a result, the authors, with support from the GMDSI, endeavored to build on the positive aspects of using Jupyter Notebooks and explore alternative teaching environments instead of week-long classes. The first major change was to add sufficient narration and explanation to the notebooks to improve possibilities for self-study. The initial design through in-person instruction was to have the notebooks serve as illustrations to assist in a narrative discussion, so bolstering of the explanatory text was necessary to help them stand alone. The next change was to refactor the organization from a strictly linear progression to the current three-part organization discussed below. This led to a hybrid model of self-study punctuated by discussion and background lectures online. 
 
+# Contents and Instructional Design
+The tutorial notebooks are structured into three main parts:
+
+## Part 0: Introductory Background
+
+> After completing this module, learners should be able to:
+>   - Describe the structure and components of the synthetic Freyberg model.
+>   - Apply the `pyemu` Python package to interface with PEST/PEST++.
+>   - Discuss the mathematical concepts that underpin the PEST/PEST++ software.
+> 
+
+Part 0 serves as the foundation, providing essential background material. Learners are encouraged to reference notebooks in Part 0 to polish their understanding of concepts they encounter in Parts 1 and 2. Part 0 is not intended to be a comprehensive resource for all background material, but rather to establish a solid understanding of the basics. The explanations of mathematical concepts are intended to be accessible through visualization and descriptions related to everyday concepts and modelling concepts. 
+
+Each notebook in Part 0 is standalone and covers a unique topic. These include:
+
+  - Introduction to a synthetic model known as the "Freyberg" model [@freyberg]. This model is used as a consistent example throughout the tutorial exercises, allowing learners to apply concepts in a practical context.
+  - An introduction to the `pyemu` Python package that is used to complement and interface with PEST/PEST++.
+  - Explanation of fundamental mathematical concepts that are relevant and will be encountered throughout the tutorial notebooks.
+
+Pre-requisites for Part 0 include a basic understanding of Python, Jupyter Notebooks, and MODFLOW 6 [@Langevin_MODFLOW_6_Modular_2022]. Familiarity with git is a bonus but not fundamental.
+
+
+
+## Part 1: Introduction to ``PEST`` and the Gauss-Levenberg Marquardt Approach
+
+> After completing this module, learners should be able to:
+> 
+> - Describe the GLM approach to parameter estimation.
+> - Explain non-uniqueness, identifiability and equifinality in the context of decision-support > modelling.
+> - Construct a ``PEST`` control file and interface files.
+> - Understand the effect of parameterization schemes on model predictive ability.
+> - Explain the difference between linear and non-linear uncertainty analysis approaches and > the value of each.
+
+Part 1 focuses on the Gauss-Levenberg Marquardt (GLM) approach to parameter estimation and associated uncertainty analysis in a groundwater modelling context. This was the foundation of the ``PEST`` software for multiple decades and the theory continues to resonate through newer techniques.
+
+Part 1 is designed to be accessible without strict sequential dependencies. Learners have the flexibility to explore its contents in any order that suits their preferences or needs. These include:
+
+  - Introduction to concepts such as non-uniqueness, identifiability, and equifinality.
+  - Introduction to the ``PEST`` control file and the PEST/PEST++ interface.
[…]
