### 2024-09-26 #240 @rhugman
> Thanks for the clarification @rhugman. I had been looking at the website.
>
> Thank you for revising the paper. I apologize for this, but I think it would be worth putting the "Content" section back in the paper. It will make the paper longer, but something is lost now that it is removed.
>
> Once this is in place, we can move ahead with publication. Thank you for your patience!!

No worries! Moved it back into the paper: [here](https://github.com/gmdsi/GMDSI_notebooks/commit/70b258794f399d828363b7cd46ea2c7f4246ef04)
