### 2024-09-26 7d8acd8
Update README.md; moving contents back to paper.md
