### 2024-09-29 #240 @kls2177
Thanks @rhugman! The next step in the publication process will to make sure that your archive doi is current. I see that you have the following [archive](https://www.sciencebase.gov/catalog/item/631a06b1d34e36012efa64ad). Please be sure that it is up-to-date. Thanks!
