### 2024-09-30 #240 @rhugman
> Thanks @rhugman! The next step in the publication process will to make sure that your archive doi is current. I see that you have the following [archive](https://www.sciencebase.gov/catalog/item/631a06b1d34e36012efa64ad). Please be sure that it is up-to-date. Thanks!

Thanks, I confirm that the [archive](https://doi.org/10.5066/P901F9G0) is up to date.
