### 2024-10-02 #240 @kls2177
I just updated the branch with the revised paper.md document. I will generate the pdf again.
