+++ paper/paper.bib
@@ -14,7 +14,8 @@
   volume={85},
   pages={217--228},
   year={2016},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.envsoft.2016.08.017}
 }
 
 @article{White_A_python_framework_2016,
@@ -119,7 +120,7 @@ address = "San Diego",
 year = "2015",
 pages = "443 - 491",
 isbn = "978-0-12-058103-0",
-doi = "http://dx.doi.org/10.1016/B978-0-08-091638-5.00010-9"
+doi = "10.1016/B978-0-08-091638-5.00010-9"
 }
 
 
@@ -199,7 +200,7 @@ doi = "http://dx.doi.org/10.1016/B978-0-08-091638-5.00010-9"
   pages = {105952},
   year = {2024},
   issn = {1364-8152},
-  doi = {https://doi.org/10.1016/j.envsoft.2024.105952},
+  doi = {10.1016/j.envsoft.2024.105952},
   url = {https://www.sciencedirect.com/science/article/pii/S1364815224000136},
   author = {Michael N. Fienen and Nicholas T. Corson-Dosch and Kalle L. Jahn and Jeremy T. White},
   keywords = {Constrained optimization, Groundwater remediation, Numerical modeling, Environmental modeling},
@@ -236,7 +237,7 @@ doi = "http://dx.doi.org/10.1016/B978-0-08-091638-5.00010-9"
   volume={10},          
   year={2022},      	  
   url={https://www.frontiersin.org/articles/10.3389/feart.2022.904271},       
-  dio={10.3389/feart.2022.904271},      	
+  doi={10.3389/feart.2022.904271},      	
   ISSN={2296-6463},     
   abstract={The Vale do Lobo sector of the Campina de Faro aquifer system in the Algarve (Portugal) is at risk of seawater intrusion. Managed Aquifer Recharge (MAR) is being considered to avoid groundwater quality deterioration. Numerical modelling was undertaken to assess the feasibility of several proposed MAR schemes. Although some data is available, many aspects of system behaviour are not well understood or measured. We demonstrate the use of a structurally simple but parametrically complex model for decision-making in a coastal aquifer. Modelling was designed to facilitate uncertainty reduction through data assimilation where possible, whilst acknowledging that which remains unknown elsewhere. Open-source software was employed throughout, and the workflow was scripted (reproducible). The model was designed to be fast-running (rapid) and numerically stable to facilitate data assimilation and represent prediction-pertinent uncertainty (robust). Omitting physical processes and structural detail constrains the type of predictions that can be made. This was addressed by assessing the effectiveness of MAR at maintaining the fresh-seawater interface (approximated using the Ghyben-Herzberg relationship) below specified thresholds. This enabled the use of a constant-density model, rather than attempting to explicitly simulating the interaction between fresh and seawater. Although predictive uncertainty may be increased, it is outweighed by the ability to extract information from the available data. Results show that, due to the limit on water availability and the continued groundwater extraction at unsustainable rates, only limited improvements in hydraulic heads can be achieved with the proposed MAR schemes. This is an important finding for decision-makers, as it indicates that a considerable reduction in extraction in addition to MAR will be required. Our approach identified these limitations, avoiding the need for further data collection, and demonstrating the value of purposeful model design.}
   }
