### 2024-10-02 #240 @rhugman
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> ❌ MISSING DOIs
>
> - 10.1016/j.envsoft.2016.08.017 may be a valid DOI for title: A python framework for environmental model uncerta...
> - 10.3389/feart.2022.904271 may be a valid DOI for title: Decision-Support Groundwater Modelling of Managed ...
>
> ❌ INVALID DOIs
>
> - http://dx.doi.org/10.1016/B978-0-08-091638-5.00010-9 is INVALID because of 'https://doi.org/' prefix
> - https://doi.org/10.1016/j.envsoft.2024.105952 is INVALID because of 'https://doi.org/' prefix
> ```

Addressed [here](https://github.com/gmdsi/GMDSI_notebooks/commit/83b9fe088efb9b671ea5756a07dc010c14ab636e).
