### 2024-10-02 #240 @kls2177
Hi @rhugman, the generated pdf is 8 pages long, which is much longer than typical JOSE papers. Although I feel the Content section should stay, some revision is required.  I have made some suggested edits in a word doc (attached) in an attempt to consolidate and reduce the word count. Please let me know if you feel that these edits still capture the essence of what you intend for the paper.

I felt that the "Statement of Need" and "Story of the Project" contained some repetitive and/or complementary material that could be consolidated. I have also removed the learning outcomes from the paper although I these should still be included in the repository. I have also made a few other smaller edits in an attempt to reduce the word count.

Also, please check the references (see reference check above). Thanks for your patience!

[groundwater_paper_suggested_edits.docx](https://github.com/user-attachments/files/17232365/groundwater_paper_suggested_edits.docx)
