### 2024-10-02 0918845
merging @kls2177 edits to paper
