### 2024-10-02 ee8c32e
returning learning objectives to reamde.md
