+++ README.md
@@ -41,8 +41,22 @@ After completing this module, learners should be able to:
 **To run the Part 0 notebooks**
 The part 0 notebooks can be run in any order and do not depend on each other. We recommend, in particular, the first 3 topics as background for the rest of the tutorials, but the remaining part 0 notebooks stand alone for more theoretical basis and a reference for topics covered later.
 
+> After completing this module, learners should be able to:
+>   - Describe the structure and components of the synthetic Freyberg model.
+>   - Apply the `pyemu` Python package to interface with PEST/PEST++.
+>   - Discuss the mathematical concepts that underpin the PEST/PEST++ software.
+> 
+
 **To run the Part 1 notebooks**
 The part 1 notebooks can also be run in any order and do not depend on each other.
+
+> After completing this module, learners should be able to:
+> 
+> - Describe the GLM approach to parameter estimation.
+> - Explain non-uniqueness, identifiability and equifinality in the context of decision-support > modelling.
+> - Construct a ``PEST`` control file and interface files.
+> - Understand the effect of parameterization schemes on model predictive ability.
+> - Explain the difference between linear and non-linear uncertainty analysis approaches and > the value of each.
 
 **To run the Part 2 notebooks**
 
@@ -77,6 +91,12 @@ PEST++MOU:
  2. part2_09_mou/freyberg_mou_2.ipynb
  
 IMPORTANT NOTE: When running some of these notebooks, PEST++ undertakes parallel model runs for computational efficiency. These runs are coordinated using internet communication protocols (e.g. TCP/IP). As a result, it is common (but not universal) that the first time a set of parallel runs is launched, a Windows Defender (or similar popup on other operating systems) box appears warning the user. This box can simply be closed by clicking "cancel" and then the model runs should proceed without issue.
+
+> After completing this module, learners should be able to:
+> 
+> - Construct a high-dimensional ``PEST`` control file and interface files.
+> - Implement GLM and ensemble-based parameter estimation and uncertainty analysis workflows > using ``PEST++`` and associated tools.
+> - Apply management optimization and sequential data assimilation techniques to > decision-support groundwater modelling problems.
 
 ## Installation Instructions
 
