### 2024-10-02 #240 @rhugman
No worries @kls2177. Thanks for the edits.

Ive added them in here [here](https://github.com/gmdsi/GMDSI_notebooks/commit/091884526965c945f9d3b25f9a4ce15891cd276f).

The pdf is still ~6 pages. If you feel it needs to be shortened further, we could remove Fig2 and the paragraph that discusses it...that should gain 1 page...
