+++ paper/paper.md
@@ -72,7 +72,7 @@ The Groundwater Modelling Decision Support Initiative ([GMDSI](https://gmdsi.org
 The roots of the materials making up the tutorial notebooks were from a traditional, week-long classroom course curriculum developed for internal training at the US Geological Survey (USGS) by a subset of the authors of this paper. After three iterations of teaching the in-person class, the authors, with support from the GMDSI, endeavored to build on the positive aspects of using Jupyter Notebooks and explore alternative teaching environments. The first major change was to add sufficient narration and explanation to the notebooks to improve possibilities for self-study. The next change was to reorganize the content from a strictly linear progression to a three-part structure. This led to a hybrid model of self-study accompanied by discussion and background lectures online. 
 
 # Statement of Need
-Many groundwater modelers typically rely on Graphical User Interfaces (GUIs) for their modelling needs. However, each GUI has its unique characteristics and varying degrees of compatibility with external software like PEST and PEST++. Creating educational materials for these GUIs would necessitate tailoring content to each GUI's specific features, obtaining cooperation from the GUI developers themselves and potentially lagging behind the latest developments. Many GUIs are commercial products as well which limits accessibility.
+Many groundwater modelers typically rely on Graphical User Interfaces (GUIs) for their modelling needs. However, each GUI has its unique characteristics and varying degrees of compatibility with external software like PEST [@pest] and PEST++ [@whitepestpp]. Creating educational materials for these GUIs would necessitate tailoring content to each GUI's specific features, obtaining cooperation from the GUI developers themselves and potentially lagging behind the latest developments. Many GUIs are commercial products as well which limits accessibility.
 
 Decision-support modelling often demands capabilities that surpass what current GUIs can offer. Thus, the use of Python for environmental modelling has increased in recent years, due to its open-source nature, user-friendly syntax, and extensive scientific libraries. Python-based tools have been developed to facilitate UQ and PE analyses, such as pyemu [@White_A_python_framework_2016; @white2021towards]. pyemu is a Python package that provides a framework for implementing UQ and PE analyses with PEST and PEST++. It offers a range of capabilities, including parameter estimation, uncertainty analysis, and management optimization. Although initially designed for groundwater modelling, pyemu's methodologies are versatile and can be applied to diverse numerical environmental models, Anecdotally, we have seen that more modelers are turning to Python packages like FloPy [@flopy] and pyemu [@whitepyemu] for model and PEST++ setup. Unfortunately, the adoption of this approach is hindered by a steep learning curve primarily due to the scarcity of user-friendly training materials. 
 
