### 2024-10-03 #240 @editorialbot
:wave: @openjournals/jose-eics, this paper is ready to be accepted and published.
