### 2024-10-10 #240 @labarba
Folks, would you kindly double check everything according to our post-review checklist above?
I noticed for example that the repository doesn't show the v1.0 [tagged release](https://github.com/gmdsi/GMDSI_notebooks/tags).
