### 2024-10-10 #240 @labarba
I see that the archive is at https://doi.org/10.5066/P901F9G0 on the ScienceBase catalog USGS repository. There, the title is "Parameter Estimation, Uncertainty Analysis, and Optimization with the PEST++ Family of codes: Tutorial Jupyter Notebooks"—but the title in the paper is "Self-Guided Decision Support Groundwater Modelling with Python." The author list also doesn't match.
