### 2024-10-10 #240 @labarba
It looks like the ScienceBase landing page for the deposited item doesn't list a license. That's unfortunate…
