### 2024-10-15 #240 @rhugman
> I see that the archive is at https://doi.org/10.5066/P901F9G0 on the ScienceBase catalog USGS repository. There, the title is "Parameter Estimation, Uncertainty Analysis, and Optimization with the PEST++ Family of codes: Tutorial Jupyter Notebooks"—but the title in the paper is "Self-Guided Decision Support Groundwater Modelling with Python." The author list also doesn't match.

Apologies @labarba and @kls2177, my misunderstanding.

Please see updated archive with current release, including license matching the repo: [https://doi.org/10.5281/zenodo.13933751](https://doi.org/10.5281/zenodo.13933751)

The repo has been updated with a [v1.0 release.](https://github.com/gmdsi/GMDSI_notebooks/releases/tag/v1.0)

Please let me know if I've missed anything else.

Cheers,
-R
