### 2024-10-15 f9e4e14
Update ci.yml to python 3.11
