+++ .github/workflows/ci.yml
@@ -21,7 +21,7 @@ jobs:
       fail-fast: false
       matrix:
         os: ["windows-latest", "ubuntu-latest", "macos-latest"]
-        python-version: ["3.9"]
+        python-version: ["3.11"]
         test-path: [
               "nb_tests.py"
                   ]      
