+++ .github/workflows/ci.yml
@@ -34,7 +34,7 @@ jobs:
       - name: Setup Mambaforge
         uses: conda-incubator/setup-miniconda@v2
         with:
-            miniforge-variant: Mambaforge
+            miniforge-variant: Miniforge
             miniforge-version: latest
             activate-environment: gmdsitut
             use-mamba: true
