### 2024-10-15 f12a1fa
Update ci.yml
tryna fix actions
