### 2024-10-15 aa574d8
Update ci.yml again
still miniforge
