### 2024-12-24 #240 @rhugman
> Congratulations, @rhugman, your JOSE paper is published! 🚀
>
>
>
> Huge thanks to our Editor: @kls2177 and Reviewers: @codyalbertross, @incsanchezro — your contributions are what makes this adventure possible 🙏

Many thanks @labarba, @kls2177! And for the reviews @incsanchezro and @codyalbertross.
